README: fix demo URL markdown
Previously the URL demo was in markdown
of the form:
[https://www.mapbox.com/bites/00223/]()

which when clicked would instead lead to non-existent URL
https://github.com/mollymerp/geofencing/blob/master

instead of going to the demo.

This fixes it by removing thee fix is to remove the markdown
since it is already a well schemeified URL so the Markdown
viewer will render it as such and clickable.